Converted to f64

+++ src/primitives/intersect.rs
@@ -1,5 +1,5 @@
 use crate::ray::{Ray, Hit};
 
 pub trait Intersect: Send + Sync + std::fmt::Debug {
-    fn intersection(&self, ray: &Ray, t_min: f32, t_max: f32) -> Option<Hit>;
+    fn intersection(&self, ray: &Ray, t_min: f64, t_max: f64) -> Option<Hit>;
 }
